+++ doc/Changelog.md
@@ -5,6 +5,8 @@
 **Not yet released**
 
 * Added [`three`](Rule-Reference.md#three-c-).
+* Fixed the UTF-8 decoder to no longer accept UTF-16 surrogates.
+* Fixed `pegtl/contrib/unescape.hh` to no longer accept unmatched surrogates.
 
 ## 2.5.2